+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatPasswordFieldUI.java
@@ -20,14 +20,20 @@ import static com.formdev.flatlaf.util.UIScale.scale;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics;
+import java.awt.Toolkit;
 import java.awt.event.FocusListener;
+import java.awt.event.KeyAdapter;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
 import java.beans.PropertyChangeEvent;
+import javax.swing.Icon;
 import javax.swing.JComponent;
 import javax.swing.LookAndFeel;
 import javax.swing.UIManager;
 import javax.swing.plaf.ComponentUI;
 import javax.swing.plaf.basic.BasicPasswordFieldUI;
 import javax.swing.text.Caret;
+import javax.swing.text.JTextComponent;
 import com.formdev.flatlaf.FlatClientProperties;
 
 /**
@@ -56,6 +62,7 @@ import com.formdev.flatlaf.FlatClientProperties;
  * @uiDefault Component.minimumWidth				int
  * @uiDefault Component.isIntelliJTheme				boolean
  * @uiDefault PasswordField.placeholderForeground	Color
+ * @uiDefault PasswordField.capsLockIcon			Icon
  * @uiDefault TextComponent.selectAllOnFocusPolicy	String	never, once (default) or always
  *
  * @author Karl Tauber
@@ -68,8 +75,10 @@ public class FlatPasswordFieldUI
 	protected int minimumWidth;
 	protected boolean isIntelliJTheme;
 	protected Color placeholderForeground;
+	protected Icon capsLockIcon;
 
 	private FocusListener focusListener;
+	private KeyListener capsLockListener;
 
 	public static ComponentUI createUI( JComponent c ) {
 		return new FlatPasswordFieldUI();
@@ -85,6 +94,7 @@ public class FlatPasswordFieldUI
 		minimumWidth = UIManager.getInt( "Component.minimumWidth" );
 		isIntelliJTheme = UIManager.getBoolean( "Component.isIntelliJTheme" );
 		placeholderForeground = UIManager.getColor( prefix + ".placeholderForeground" );
+		capsLockIcon = UIManager.getIcon( "PasswordField.capsLockIcon" );
 
 		LookAndFeel.installProperty( getComponent(), "opaque", focusWidth == 0 );
 
@@ -96,6 +106,7 @@ public class FlatPasswordFieldUI
 		super.uninstallDefaults();
 
 		placeholderForeground = null;
+		capsLockIcon = null;
 
 		MigLayoutVisualPadding.uninstall( getComponent() );
 	}
@@ -105,7 +116,23 @@ public class FlatPasswordFieldUI
 		super.installListeners();
 
 		focusListener = new FlatUIUtils.RepaintFocusListener( getComponent() );
+		capsLockListener = new KeyAdapter() {
+			@Override
+			public void keyPressed( KeyEvent e ) {
+				repaint( e );
+			}
+			@Override
+			public void keyReleased( KeyEvent e ) {
+				repaint( e );
+			}
+			private void repaint( KeyEvent e ) {
+				if( e.getKeyCode() == KeyEvent.VK_CAPS_LOCK )
+					e.getComponent().repaint();
+			}
+		};
+
 		getComponent().addFocusListener( focusListener );
+		getComponent().addKeyListener( capsLockListener );
 	}
 
 	@Override
@@ -113,7 +140,9 @@ public class FlatPasswordFieldUI
 		super.uninstallListeners();
 
 		getComponent().removeFocusListener( focusListener );
+		getComponent().removeKeyListener( capsLockListener );
 		focusListener = null;
+		capsLockListener = null;
 	}
 
 	@Override
@@ -133,7 +162,19 @@ public class FlatPasswordFieldUI
 	protected void paintSafely( Graphics g ) {
 		FlatTextFieldUI.paintBackground( g, getComponent(), focusWidth, arc, isIntelliJTheme );
 		FlatTextFieldUI.paintPlaceholder( g, getComponent(), placeholderForeground );
+		paintCapsLock( g );
 		super.paintSafely( g );
+	}
+
+	protected void paintCapsLock( Graphics g ) {
+		JTextComponent c = getComponent();
+		if( !c.isFocusOwner() ||
+			!Toolkit.getDefaultToolkit().getLockingKeyState( KeyEvent.VK_CAPS_LOCK ) )
+		  return;
+
+		int y = (c.getHeight() - capsLockIcon.getIconHeight()) / 2;
+		int x = c.getWidth() - capsLockIcon.getIconWidth() - y;
+		capsLockIcon.paintIcon( c, g, x, y );
 	}
 
 	@Override