+++ CHANGELOG.md
@@ -3,6 +3,7 @@ FlatLaf Change Log
 
 ## Unreleased
 
+- PasswordField: Warn about enabled Caps Lock.
 - TextField, FormattedTextField and PasswordField: Support round borders (see UI
   default value `TextComponent.arc`). (issue #65)
 - IntelliJ Themes: Added Gradianto themes to demo.